Created new SignalKSource class
Moved functionality of SignalK specific methods to its own class.
