+++ src/transforms/transform.cpp
@@ -8,7 +8,7 @@
 
 std::set<TransformBase*> TransformBase::transforms;
 
-TransformBase::TransformBase(String sk_path, String id, String schema)
-    : Configurable{id, schema}, sk_path{sk_path}, MustEnable(5) {
+TransformBase::TransformBase(String sk_path, String id, String schema) :
+    Configurable{id, schema}, SignalKSource(sk_path), MustEnable(5) {
   transforms.insert(this);
 }
